Added Trhythm::isValid() when rhythm is different than none, fixed method resolving duration

+++ src/libs/core/music/tmeter.cpp
@@ -53,7 +53,7 @@ int Tmeter::lower() const {
 }
 
 
-quint16 Tmeter::duration() const {
+int Tmeter::duration() const {
   if (NoMeter)
     return 0;
   else
